+++ maven/src/main/java/plushie_tycoon/global/GlobalMain.java
@@ -4,10 +4,8 @@ import plushie_tycoon.global.geGlobal.GEGlobal;
 
 public class GlobalMain {
     public static void main(String[] args) throws Exception {
-        int publicPortno = 50002;
-        int adminPortno = 50003;
-        GEGlobal ge = new GEGlobal(publicPortno, adminPortno);
-        ge.runGlobalServer();
-        ge.runAdminServer();
+        int portno = 50002;
+        GEGlobal ge = new GEGlobal(portno);
+        ge.runServices();
     }
 }
